ml_engine/requirements.txt
PyPDF2>=3.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.13.0
numpy>=1.24.0
pandas>=2.0.0
jupyter>=1.0.0
jupyterlab>=4.0.0
ipython>=8.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
